refactor: move shared ui state into ui state package
